refactor: extract spec parsing helpers
